reorg test files
